+++ ms-azure-api/src/MSAzureAPI.hs
@@ -1,0 +1,1 @@
+module MSAzureAPI where