+++ ms-azure-api/src/MSAzureAPI/Internal/Common.hs
@@ -1,0 +1,78 @@
+{-# LANGUAGE DataKinds #-}
+{-# options_ghc -Wno-unused-imports #-}
+-- | Common functions for the MS Azure API
+--
+module MSAzureAPI.Internal.Common where
+
+import Control.Monad.IO.Class (MonadIO(..))
+import Data.Proxy (Proxy)
+import GHC.Generics (Generic(..))
+
+import Data.List (sort, sortBy, stripPrefix, uncons)
+import Data.Maybe (listToMaybe, fromMaybe)
+-- import Data.Ord (comparing)
+import Data.Char (toLower)
+
+-- aeson
+import qualified Data.Aeson as A (ToJSON(..), FromJSON(..), genericParseJSON, defaultOptions, Options(..), withObject, withText, (.:), (.:?), object, (.=), Key, Value, camelTo2)
+-- bytestring
+import qualified Data.ByteString as BS (ByteString)
+import qualified Data.ByteString.Char8 as BS8 (pack, unpack)
+import qualified Data.ByteString.Lazy as LBS (ByteString)
+import qualified Data.ByteString.Lazy.Char8 as LBS8 (pack, unpack, putStrLn)
+-- hoauth2
+import Network.OAuth.OAuth2 (OAuth2Token(..))
+import Network.OAuth.OAuth2.Internal (AccessToken(..), ExchangeToken(..), RefreshToken(..), OAuth2Error, IdToken(..))
+-- modern-uri
+-- import Text.URI (URI, mkURI)
+-- req
+import Network.HTTP.Req (Req, runReq, HttpException(..), defaultHttpConfig, req, Option, (=:), GET(..), POST(..), Url, Scheme(..), useHttpsURI, https, (/:), ReqBodyJson(..), NoReqBody(..), oAuth2Bearer, HttpResponse(..), jsonResponse, JsonResponse, lbsResponse, LbsResponse, bsResponse, BsResponse, responseBody)
+-- text
+import Data.Text (Text, pack, unpack)
+
+
+
+-- | Data plane e.g for FileREST API
+
+-- | Control plane
+msAzureManagementReqConfig :: AccessToken -> [Text] -> (Url 'Https, Option 'Https)
+msAzureManagementReqConfig (AccessToken ttok) uriRest = (url, os)
+  where
+    url = (https "management.azure.com") //: uriRest
+    os = oAuth2Bearer $ BS8.pack (unpack ttok)
+
+(//:) :: Url scheme -> [Text] -> Url scheme
+(//:) = foldl (/:)
+
+
+-- * aeson
+
+-- | a collection of items with key @value@
+data Collection a = Collection {
+  cValue :: [a]
+                               } deriving (Eq, Show, Generic)
+instance A.FromJSON a => A.FromJSON (Collection a) where
+  parseJSON = A.genericParseJSON (aesonOptions "c")
+
+-- | drop the prefix and lowercase first character
+--
+-- e.g. @userDisplayName@ @->@ @displayName@
+aesonOptions :: String -- ^ record prefix
+             -> A.Options
+aesonOptions pfx = A.defaultOptions { A.fieldLabelModifier = recordName pfx }
+
+-- | drop the prefix and lowercase first character
+recordName :: String -- ^ record name prefix
+           -> String -- ^ JSON field name
+           -> String
+recordName pf str = case uncons $ dropPrefix pf str of
+  Just (c, cs) -> toLower c : cs
+  _ -> error "record name cannot be empty"
+
+-- | Drops the given prefix from a list.
+--   It returns the original sequence if the sequence doesn't start with the given prefix.
+--
+-- > dropPrefix "Mr. " "Mr. Men" == "Men"
+-- > dropPrefix "Mr. " "Dr. Men" == "Dr. Men"
+dropPrefix :: Eq a => [a] -> [a] -> [a]
+dropPrefix a b = fromMaybe b $ stripPrefix a b